Move nex-gen support input out of package
