+++ temporalio/nexus/system/workflow_service/_support/__init__.py
@@ -2,4 +2,4 @@
 
 from __future__ import annotations
 
-from .temporal_model_converters import *  # noqa: F401,F403
+from .nex_gen_support import *  # noqa: F401,F403
